Add RoleClaimConfiguration for RoleClaims table
This commit introduces a new class `RoleClaimConfiguration` that implements the `IEntityTypeConfiguration<IdentityRoleClaim<string>>` interface. It configures the `RoleClaims` table in the database, specifying the table name and setting the primary key to the `Id` property of `IdentityRoleClaim<string>`. Necessary using directives for `Microsoft.AspNetCore.Identity` and `Microsoft.EntityFrameworkCore.Metadata.Builders` are also included.