Add UserLoginDtoValidator for login data validation
Introduces the `UserLoginDtoValidator` class in the
`Cinema.BLL.Validators` namespace, extending
`AbstractValidator<UserLoginDto>`. This class validates
user login data by ensuring the `Email` field is not
empty and is in a valid format, and that the `Password`
field is also not empty. Necessary using directives for
`Cinema.BLL.DTOs.Users` and `FluentValidation` have
been added.