Add ITmDbService interface for movie operations
This commit introduces the `ITmDbService` interface in the
`Cinema.BLL.Interfaces.Services` namespace. It defines two
asynchronous methods: `SearchMoviesAsync` for retrieving
a list of `MovieDto` objects based on a search query,
and `GetMovieForCreateAsync` for fetching a `MovieCreateDto`
object using a specified `tmDbId`. The necessary DTOs are
imported from the `Cinema.BLL.DTOs.Movies` namespace.